Modificado para uma melhor integridade no SQL

+++ src/main/java/dev_marcelo/maNotes/dto/FundosValoresDto.java
@@ -8,5 +8,5 @@ import lombok.Setter;
 @Getter @Setter @AllArgsConstructor @NoArgsConstructor
 public class FundosValoresDto {
     private String origemDoFundo;
-    private Float valorRecebido;
+    private Double valorRecebido;
 }
